+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This changelog follows the patterns described here: https://keepachangelog.com/e
 Subheadings to categorize changes are `added, changed, deprecated, removed, fixed, security`.
 
 ## Unreleased
+### added
+- Configurable character of PaginationEllipsis.
+
 ### fixed
 - ImageSize should store `is-square` instead of `is-Square` to html class attribute.